feat(services): extract the scheduler into @mulmoclaude/scheduler
Sixth and final packages/services/* package — the cron tick engine + the
persistence/catch-up adapter shared by MulmoClaude and MulmoTerminal, so a
MulmoTerminal-alone run no longer silently loses scheduling (journal / feeds /
user-cron). Closes the headless-backend extraction.

Two host-agnostic modules move into the package:
- task-manager: the generic dependency-ordered tick engine (createTaskManager,
  ITaskManager, TaskDefinition, TaskSchedule, …). Its only host coupling was the
  logger, now injected via TaskManagerOptions.log (defaults to noop); time constants
  inlined.
- adapter: the thin binding over @receptron/task-scheduler — catch-up plan on boot,
  per-tick window accounting, best-effort state + log persistence. Workspace root,
  the atomic writer, and the logger are injected via configureScheduler.

Host-specific pieces STAY host-side and are passed in: the system task DEFINITIONS
(journal / chat-index / feed-refresh in server/index.ts), the overrides IO
(scheduler-io / scheduler-overrides-io), registerScheduledSkills / registerUserTasks,
and the HTTP routes. The package owns no task definitions and no routes — the plan's
"the host should pick its task set."

- packages/services/scheduler: task-manager.ts + adapter.ts. Server-only, dual
  ESM+CJS, peer-dep @receptron/task-scheduler. 5 package tests (tick due-firing,
  dependsOn ordering + dep-failure skip, duplicate-id guard, initScheduler registration
  + state exposure, scheduled-run persistence to the injected workspace).
- Host server/events/task-manager/index.ts reduced to a re-export shim that injects
  MulmoClaude's logger into createTaskManager. server/events/scheduler-adapter.ts
  reduced to a re-export shim that calls configureScheduler (workspacePath +
  writeFileAtomic + the host logger) at module load.

Verification: typecheck:server/test/vue clean; task-dependencies (10), scheduler
overrides (4), schedulerHandlers route (49), plugin-runtime (27), runtime-loader (3),
bookmarks (3), recipe-book (4) — 100 scheduler/task-manager-dependent tests — all green
against the extracted package; package tests (5) green.
